// src/types/profile.ts
export enum DifficultyLevel {
  BEGINNER = 1,
  INTERMEDIATE = 2,
  ADVANCED = 3,
}

export enum DurationLevel {
  LESS_THAN_3_HOURS = 1,
  THREE_TO_SIX_HOURS = 2,
  SIX_TO_NINE_HOURS = 3,
  MORE_THAN_NINE_HOURS = 4,
}

export enum EnvironmentLevel {
  OCEAN = 1,
  MOUNTAIN = 2,
  DOESNT_MATTER = 3,
}